requirements.txt
sentence-transformers
faiss-cpu
nltk
scikit-learn
numpy
pandas
pathway
google-generativeai
tqdm
python-dotenv
